添加组件历史、历史文件接口，修改跨域请求获取token失败的问题

+++ src/main/java/com/rengu/operationsmanagementsuitev3/Service/ComponentFileService.java
@@ -168,7 +168,7 @@ public class ComponentFileService {
             componentFileRepository.deleteById(componentFileEntity.getId());
         } else {
             // 是文件，检查是否需要删除实际文件
-            if (!hasComponentFileByFile(componentFileEntity.getFileEntity())) {
+            if (!hasComponentFileByFile(componentFileEntity.getFileEntity()) && !componentFileHistoryService.hasComponentFileHistoryByFile(componentFileEntity.getFileEntity())) {
                 fileService.deleteFileById(componentFileEntity.getFileEntity().getId());
             }
             componentFileRepository.deleteById(componentFileEntity.getId());
